pyproject.toml
[tool.poetry]
name = "league-backend"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2"
django-configurations = "^2.2"
dj-database-url = "^0.5.0"
djangorestframework = "^3.12.4"
django-filter = "^2.4.0"
django-extensions = "^3.1.2"
django-storages = "^1.11.1"
Pillow = "^8.2.0"
django-cors-headers = "^3.7.0"
django-querycount = "^0.7.0"
djangorestframework-simplejwt = "^4.7.2"
psycopg2-binary = "^2.9.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"
